*** empty log message ***
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@684 7b491191-dbf0-0310-aff6-d879d4d69008

+++ user.c
@@ -16,6 +16,8 @@
 #include <crypt.h>
 #include <time.h>
 #include <stdarg.h>
+#include <sys/types.h>
+#include <unistd.h>
 
 char *configFile = "/etc/dbmail.conf";
 
@@ -559,9 +561,9 @@ int do_show(char *name)
       quiet_printf("User ID         : %llu\n", userid);
       quiet_printf("Username        : %s\n", auth_get_userid(&userid));
       quiet_printf("Client ID       : %llu\n",cid);
-      quiet_printf("Max. mailboxsize: %.02lf MB\n",(double)quotum/(1024.0*1024.0));
-      quiet_printf("Quotum used     : %.02lf MB (%.01lf%%)\n", (double)quotumused/(1024.0*1024.0), 
-		   (100.0 * quotumused)/quotum);
+      quiet_printf("Max. mailboxsize: %.02f MB\n",(double)quotum/(1024.0*1024.0));
+      quiet_printf("Quotum used     : %.02f MB (%.01f%%)\n", (double)quotumused/(1024.0*1024.0), 
+		   (100.0 * quotumused)/(double)quotum);
       quiet_printf("\n");
 
       quiet_printf("Aliases:\n");
